[BEAM-77] Move hadoop contrib as hdfs IO
