+++ ios/pom.xml
@@ -35,6 +35,7 @@
 
   <modules>
     <module>bigtable</module>
+    <module>hdfs</module>
   </modules>
 
 </project>
